requirements.txt
requests>=2.32
PyYAML>=6.0
python-dotenv>=1.0
pytest>=8.0
paramiko>=4.0
psycopg2-binary>=2.9
numpy>=2.0
pillow>=10.0
